% Auto-generated: do not edit by hand
\name{pvIcon}

\alias{pvIcon}

\title{PvIcon component}

\description{
pv-icon is a component that renders an svg icon.
}

\usage{
pvIcon(id=NULL, className=NULL, href=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. The className could be pv-icon, pv-icon-20, etc.}

\item{href}{Character. The href is the ID of the pv-icon.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

